pyproject.toml
[project]
name = "adk-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.15.1",
    "litellm>=1.77.5",
    "pillow>=11.3.0",
]
